Remove hard assertion requiring .pyi stub files to exist, which breaks
  builds with protobuf < 21.0 or when ONNX_GEN_PB_TYPE_STUBS is disabled.

  Downstream build fails with:
    AssertionError: Bug: No generated python stubs found
  when `-DONNX_GEN_PB_TYPE_STUBS=OFF`.

  This affects:
  - RHEL 9 (protobuf 3.19.6)
  - Hermetic build environments with controlled dependencies
  - Minimal builds with stub generation disabled

  Python stub files (.pyi) are type hints for static analysis tools like
  mypy, not runtime requirements. They can safely be optional per PEP 561.

  Changes:
  - Remove assertion that .pyi files must exist
  - Add comment explaining why stubs are optional
  - Stub files still copied when generated (no behavior change)

  Fixes: Build failures with protobuf < 21.0
  Tested: RHEL 9 with protobuf 3.19.6

+++ onnx/defs/generator/old.cc
@@ -79,11 +79,9 @@ ONNX_OPERATOR_SET_SCHEMA(
               FunctionBuilder builder(functionProto);
               builder
                   .Add(
-                      "X_random = RandomUniformLike <low = 0.0, high = 1.0, seed = @seed> (input)",
-                      "dtype",
-                      int64_t(input_type))
+                      "X_random = RandomUniformLike <low = 0.0, high = 1.0, seed = @seed> (input)", "dtype", input_type)
                   .Add("X_greater = Greater (X_random, input)")
-                  .Add("output = Cast (X_greater)", "to", int64_t(dtype));
+                  .Add("output = Cast (X_greater)", "to", dtype);
               schema.BuildFunction(functionProto);
               return true;
             }));
